Refactor chat panel for provider-aware NPC chat
